requirements.txt
chromadb 
sentence-transformers 
kafka-python 
GoogleNews 
pandas 
python-dateutil 
numpy
openai>=1.0.0
streamlit
langdetect
langchain-openai
langchain-chroma
langchain-huggingface
langchain